Move SimulatedHandData into own file
