+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSimulationSystem/SimulatedHand.cs
@@ -3,97 +3,13 @@
 
 using System;
 using System.Collections.Generic;
-using UnityEngine;
 using XRTK.Definitions.Devices;
 using XRTK.Definitions.Utilities;
 using XRTK.Interfaces.InputSystem;
+using XRTK.Providers.Controllers.Hands;
 
 namespace XRTK.Services.InputSimulationSystem
 {
-    /// <summary>
-    /// Snapshot of simulated hand data.
-    /// </summary>
-    [Serializable]
-    public class SimulatedHandData
-    {
-        private static readonly int jointCount = Enum.GetNames(typeof(TrackedHandJoint)).Length;
-
-        // Timestamp of hand data, as FileTime, e.g. DateTime.UtcNow.ToFileTime()
-        private long timestamp = 0;
-        public long Timestamp => timestamp;
-
-        [SerializeField]
-        private bool isTracked = false;
-        public bool IsTracked => isTracked;
-        [SerializeField]
-        private MixedRealityPose[] joints = new MixedRealityPose[jointCount];
-        public MixedRealityPose[] Joints => joints;
-        [SerializeField]
-        private bool isPinching = false;
-        public bool IsPinching => isPinching;
-
-        public delegate void HandJointDataGenerator(MixedRealityPose[] jointPositions);
-
-        private IMixedRealityInputSystem inputSystem = null;
-
-        /// <summary>
-        /// The active instance of the input system.
-        /// </summary>
-        private IMixedRealityInputSystem InputSystem
-        {
-            get
-            {
-                if (inputSystem == null)
-                {
-                    inputSystem = MixedRealityToolkit.GetService<IMixedRealityInputSystem>();
-                }
-                return inputSystem;
-            }
-        }
-
-        public void Copy(SimulatedHandData other)
-        {
-            timestamp = other.timestamp;
-            isTracked = other.isTracked;
-            isPinching = other.isPinching;
-            for (int i = 0; i < jointCount; ++i)
-            {
-                joints[i] = other.joints[i];
-            }
-        }
-
-        public bool Update(bool isTrackedNew, bool isPinchingNew, HandJointDataGenerator generator)
-        {
-            // TODO: DateTime.UtcNow can be quite imprecise, better use Stopwatch.GetTimestamp
-            // https://stackoverflow.com/questions/2143140/c-sharp-datetime-now-precision
-            return UpdateWithTimestamp(DateTime.UtcNow.Ticks, isTrackedNew, isPinchingNew, generator);
-        }
-
-        public bool UpdateWithTimestamp(long timestampNew, bool isTrackedNew, bool isPinchingNew, HandJointDataGenerator generator)
-        {
-            bool handDataChanged = false;
-
-            if (isTracked != isTrackedNew || isPinching != isPinchingNew)
-            {
-                isTracked = isTrackedNew;
-                isPinching = isPinchingNew;
-                handDataChanged = true;
-            }
-
-            if (timestamp != timestampNew)
-            {
-                timestamp = timestampNew;
-                if (isTracked)
-                {
-                    generator(Joints);
-                    handDataChanged = true;
-                }
-            }
-
-            return handDataChanged;
-        }
-    }
-
     public abstract class SimulatedHand : BaseHand
     {
         public abstract HandSimulationMode SimulationMode { get; }
@@ -134,7 +50,8 @@ namespace XRTK.Services.InputSimulationSystem
                 }
             }
 
-            InputSystem?.RaiseHandJointsUpdated(InputSource, ControllerHandedness, jointPoses);
+
+            MixedRealityToolkit.InputSystem.RaiseHandJointsUpdated(InputSource, ControllerHandedness, jointPoses);
 
             UpdateVelocity();
