+++ XRTK-Core/Packages/com.xrtk.core/Services/InputSimulationSystem/SimulatedHandData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8a6c3c472b8250042ab4834805d9b12b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
